// src/types/enums/ErrorCode.ts
export enum ErrorCode {
  REFRESH_TOKEN_MISSING = 'REFRESH_TOKEN_MISSING',
  REFRESH_TOKEN_INVALID = 'REFRESH_TOKEN_INVALID',
  REFRESH_TOKEN_EXPIRED = 'REFRESH_TOKEN_EXPIRED',
  REFRESH_TOKEN_MISMATCH = 'REFRESH_TOKEN_MISMATCH',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  SERVER_ERROR = 'SERVER_ERROR',
  PASSWORD_ERROR = 'PASSWORD_ERROR',
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  EMAIL_EXISTS = 'EMAIL_EXISTS',
  UNAUTHORIZED = 'UNAUTHORIZED',
  ACCOUNT_BLOCKED = 'ACCOUNT_BLOCKED',
  INVALID_OTP = 'INVALID_OTP'
} 